nixos/nixos-containers: add flake to container start script
Integrated the flake container setup into the spawn script for systemd-nspawn.
The trickiest part of this was ensuring the underlying per-container is built.
With the .conf file created, running `nixos-container update` creates all the necessary per-container structure.
We call this command at start to ensure the structure is created.
Note: This also means the flake gets updated to branch HEAD when the container is started/restarted.